ecommerce/users/serializers.py
from django.shortcuts import render

from rest_framework import serializers
from django.contrib.auth.models import User
from django.contrib.auth.hashers import make_password

class UserSerializer(serializers.ModelSerializer):
    password = serializers.CharField(write_only = True)
    password2 = serializers.CharField(write_only = True)
    class Meta:
        model = User
        fields = ['username', 'first_name', 'last_name', 'email', 'password', 'password2']
    
    def validate(request, data):
        if data["password"] != data["password2"]:
            raise serializers.ValidationError("passwords must match")
        return data
    
    def create(self, validated_data):
        password = validated_data.pop("password")
        validated_data.pop("password2")
        user = User.objects.create_user(**validated_data, password = password)
        return user